@@ -7,7 +7,6 @@ import { UploadFile } from "./class-interfaces/upload";
 import { CategoriesAPI } from "./class-interfaces/categories";
 import { TagsAPI } from "./class-interfaces/tags";
 import { UtilsAPI } from "./class-interfaces/utils";
-import { NotificationsAPI } from "./class-interfaces/event-notifications";
 import { FoodAPI } from "./class-interfaces/recipe-foods";
 import { UnitAPI } from "./class-interfaces/recipe-units";
 import { CookbookAPI } from "./class-interfaces/group-cookbooks";
@@ -23,10 +22,10 @@ import { GroupMigrationApi } from "./class-interfaces/group-migrations";
 import { GroupReportsApi } from "./class-interfaces/group-reports";
 import { ShoppingApi } from "./class-interfaces/group-shopping-lists";
 import { MultiPurposeLabelsApi } from "./class-interfaces/group-multiple-purpose-labels";
+import { GroupEventNotifierApi } from "./class-interfaces/group-event-notifier";
 import { ApiRequestInstance } from "~/types/api";
 
 class Api {
-  // private static instance: Api;
   public recipes: RecipeAPI;
   public users: UserApi;
   public groups: GroupAPI;
@@ -35,7 +34,6 @@ class Api {
   public categories: CategoriesAPI;
   public tags: TagsAPI;
   public utils: UtilsAPI;
-  public notifications: NotificationsAPI;
   public foods: FoodAPI;
   public units: UnitAPI;
   public cookbooks: CookbookAPI;
@@ -50,14 +48,10 @@ class Api {
   public tools: ToolsApi;
   public shopping: ShoppingApi;
   public multiPurposeLabels: MultiPurposeLabelsApi;
-  // Utils
+  public groupEventNotifier: GroupEventNotifierApi;
   public upload: UploadFile;
 
   constructor(requests: ApiRequestInstance) {
-    // if (Api.instance instanceof Api) {
-    //   return Api.instance;
-    // }
-
     // Recipes
     this.recipes = new RecipeAPI(requests);
     this.categories = new CategoriesAPI(requests);
@@ -84,7 +78,6 @@ class Api {
     // Admin
     this.events = new EventsAPI(requests);
     this.backups = new BackupAPI(requests);
-    this.notifications = new NotificationsAPI(requests);
 
     // Utils
     this.upload = new UploadFile(requests);
@@ -92,9 +85,9 @@ class Api {
 
     this.email = new EmailAPI(requests);
     this.bulk = new BulkActionsAPI(requests);
+    this.groupEventNotifier = new GroupEventNotifierApi(requests);
 
     Object.freeze(this);
-    // Api.instance = this;
   }
 }
 
