@@ -2,9 +2,10 @@ from functools import cached_property
 
 from sqlalchemy.orm import Session
 
-from mealie.db.models.event import Event, EventNotification
+from mealie.db.models.event import Event
 from mealie.db.models.group import Group, GroupMealPlan, ReportEntryModel, ReportModel
 from mealie.db.models.group.cookbook import CookBook
+from mealie.db.models.group.events import GroupEventNotifierModel
 from mealie.db.models.group.exports import GroupDataExportsModel
 from mealie.db.models.group.invite_tokens import GroupInviteToken
 from mealie.db.models.group.preferences import GroupPreferencesModel
@@ -24,7 +25,7 @@ from mealie.db.models.users import LongLiveToken, User
 from mealie.db.models.users.password_reset import PasswordResetModel
 from mealie.schema.cookbook.cookbook import ReadCookBook
 from mealie.schema.events import Event as EventSchema
-from mealie.schema.events import EventNotificationIn
+from mealie.schema.group.group_events import GroupEventNotifierOut
 from mealie.schema.group.group_exports import GroupDataExport
 from mealie.schema.group.group_preferences import ReadGroupPreferences
 from mealie.schema.group.group_shopping_list import ShoppingListItemOut, ShoppingListOut
@@ -117,10 +118,6 @@ class AllRepositories:
         return RepositoryGeneric(self.session, pk_id, SignUp, SignUpOut)
 
     @cached_property
-    def event_notifications(self) -> RepositoryGeneric[EventNotificationIn, EventNotification]:
-        return RepositoryGeneric(self.session, pk_id, EventNotification, EventNotificationIn)
-
-    @cached_property
     def events(self) -> RepositoryGeneric[EventSchema, Event]:
         return RepositoryGeneric(self.session, pk_id, Event, EventSchema)
 
@@ -193,3 +190,7 @@ class AllRepositories:
     @cached_property
     def group_multi_purpose_labels(self) -> RepositoryGeneric[MultiPurposeLabelOut, MultiPurposeLabel]:
         return RepositoryGeneric(self.session, pk_id, MultiPurposeLabel, MultiPurposeLabelOut)
+
+    @cached_property
+    def group_event_notifier(self) -> RepositoryGeneric[GroupEventNotifierOut, GroupEventNotifierModel]:
+        return RepositoryGeneric(self.session, pk_id, GroupEventNotifierModel, GroupEventNotifierOut)
