@@ -1,4 +1,3 @@
-import apprise
 from sqlalchemy.orm.session import Session
 
 from mealie.db.db_setup import create_session
@@ -6,43 +5,11 @@ from mealie.repos.all_repositories import get_repositories
 from mealie.schema.events import Event, EventCategory
 
 
-def test_notification(notification_url, event=None) -> bool:
-    if event is None:
-        event = Event(
-            title="Test Notification",
-            text="This is a test message from the Mealie API server",
-            category=EventCategory.general.value,
-        )
-
-    post_notifications(event, [notification_url], hard_fail=True)
-
-
-def post_notifications(event: Event, notification_urls=list[str], hard_fail=False, attachment=None):
-    asset = apprise.AppriseAsset(async_mode=False)
-    apobj = apprise.Apprise(asset=asset)
-
-    for dest in notification_urls:
-        status = apobj.add(dest)
-
-        if not status and hard_fail:
-            raise Exception("Apprise URL Add Failed")
-
-    apobj.notify(
-        body=event.text,
-        title=event.title,
-        attach=str(attachment),
-    )
-
-
-def save_event(title, text, category, session: Session, attachment=None):
+def save_event(title, text, category, session: Session):
     event = Event(title=title, text=text, category=category)
     session = session or create_session()
     db = get_repositories(session)
     db.events.create(event.dict())
-
-    notification_objects = db.event_notifications.get(match_value=True, match_key=category, limit=9999)
-    notification_urls = [x.notification_url for x in notification_objects]
-    post_notifications(event, notification_urls, attachment=attachment)
 
 
 def create_general_event(title, text, session=None):
@@ -52,8 +19,7 @@ def create_general_event(title, text, session=None):
 
 def create_recipe_event(title, text, session=None, attachment=None):
     category = EventCategory.recipe
-
-    save_event(title=title, text=text, category=category, session=session, attachment=attachment)
+    save_event(title=title, text=text, category=category, session=session)
 
 
 def create_backup_event(title, text, session=None):
